refactor(pos) :  패키지 - common 추가 및 purgomalum 관련 도메인 이동
