+++ src/main/java/kitchenpos/common/domain/model/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.infra;
+package kitchenpos.common.domain.model;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
