+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -3,7 +3,7 @@ package kitchenpos.product.application;
 import kitchenpos.menu.domain.model.Menu;
 import kitchenpos.menu.domain.model.MenuProduct;
 import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.menu.infra.PurgomalumClient;
+import kitchenpos.common.domain.model.PurgomalumClient;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
