+++ src/test/java/kitchenpos/menu/infra/FakePurgomalumClient.java
@@ -1,5 +1,7 @@
 package kitchenpos.menu.infra;
 
+import kitchenpos.common.domain.model.PurgomalumClient;
+
 import java.util.Arrays;
 import java.util.List;
