+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -4,7 +4,7 @@ import kitchenpos.menu.infra.FakePurgomalumClient;
 import kitchenpos.menu.application.InMemoryMenuRepository;
 import kitchenpos.menu.domain.model.Menu;
 import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.menu.infra.PurgomalumClient;
+import kitchenpos.common.domain.model.PurgomalumClient;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
